+++ beacon-chain/blockchain/process_block_helpers.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/feed"
 	statefeed "github.com/prysmaticlabs/prysm/v5/beacon-chain/core/feed/state"
+	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/helpers"
 	lightclient "github.com/prysmaticlabs/prysm/v5/beacon-chain/core/light-client"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/transition"
 	doublylinkedtree "github.com/prysmaticlabs/prysm/v5/beacon-chain/forkchoice/doubly-linked-tree"
@@ -138,23 +139,27 @@ func (s *Service) saveLightClientUpdate(cfg *postBlockProcessConfig) {
 	attestedRoot := cfg.roblock.Block().ParentRoot()
 	attestedBlock, err := s.getBlock(cfg.ctx, attestedRoot)
 	if err != nil {
-		log.WithError(err).Error("Could not get attested block")
+		log.WithError(err).Error("Saving light client update failed: Could not get attested block")
 		return
 	}
-	if attestedBlock == nil {
-		log.Error("Attested block is nil")
+	if attestedBlock == nil || attestedBlock.IsNil() {
+		log.Error("Saving light client update failed: Attested block is nil")
 		return
 	}
 	attestedState, err := s.cfg.StateGen.StateByRoot(cfg.ctx, attestedRoot)
-	if err != nil || attestedState == nil {
-		log.WithError(err).Error("Could not get attested state")
+	if err != nil {
+		log.WithError(err).Error("Saving light client update failed: Could not get attested state")
+		return
+	}
+	if attestedState == nil || attestedState.IsNil() {
+		log.Error("Saving light client update failed: Attested state is nil")
 		return
 	}
 
 	finalizedRoot := attestedState.FinalizedCheckpoint().Root
 	finalizedBlock, err := s.getBlock(cfg.ctx, [32]byte(finalizedRoot))
 	if err != nil {
-		log.WithError(err).Error("Could not get finalized block")
+		log.WithError(err).Error("Saving light client update failed: Could not get finalized block")
 		return
 	}
 
@@ -168,33 +173,43 @@ func (s *Service) saveLightClientUpdate(cfg *postBlockProcessConfig) {
 		finalizedBlock,
 	)
 	if err != nil {
-		log.WithError(err).Error("Could not create light client update")
+		log.WithError(err).Error("Saving light client update failed: Could not create light client update")
 		return
 	}
 
-	period := uint64(attestedState.Slot()) / (uint64(params.BeaconConfig().SlotsPerEpoch) * uint64(params.BeaconConfig().EpochsPerSyncCommitteePeriod))
+	period := helpers.SyncCommitteePeriodBySlot(attestedState.Slot())
 
 	oldUpdate, err := s.cfg.BeaconDB.LightClientUpdate(cfg.ctx, period)
 	if err != nil {
-		log.WithError(err).Error("Could not get current light client update")
+		log.WithError(err).Error("Saving light client update failed: Could not get current light client update")
 		return
 	}
+
 	if oldUpdate == nil {
 		if err := s.cfg.BeaconDB.SaveLightClientUpdate(cfg.ctx, period, update); err != nil {
-			log.WithError(err).Error("Could not save light client update")
+			log.WithError(err).Error("Saving light client update failed: Could not save light client update")
+		} else {
+			log.WithField("period", period).Debug("Saving light client update: Saved new update")
+		}
+		return
+	}
+
+	isNewUpdateBetter, err := lightclient.IsBetterUpdate(update, oldUpdate)
+	if err != nil {
+		log.WithError(err).Error("Saving light client update failed: Could not compare light client updates")
+		return
+	}
+
+	if isNewUpdateBetter {
+		if err := s.cfg.BeaconDB.SaveLightClientUpdate(cfg.ctx, period, update); err != nil {
+			log.WithError(err).Error("Saving light client update failed: Could not save light client update")
+		} else {
+			log.WithField("period", period).Debug("Saving light client update: Saved new update")
 		}
 	} else {
-		isNewUpdateBetter, err := lightclient.IsBetterUpdate(update, oldUpdate)
-		if err != nil {
-			log.WithError(err).Error("Could not compare light client updates")
-			return
-		}
-		if isNewUpdateBetter {
-			if err := s.cfg.BeaconDB.SaveLightClientUpdate(cfg.ctx, period, update); err != nil {
-				log.WithError(err).Error("Could not save light client update")
-			}
-		}
+		log.WithField("period", period).Debug("Saving light client update: New update is not better than the current one. Skipping save.")
 	}
+
 }
 
 // saveLightClientBootstrap saves a light client bootstrap for this block
@@ -207,12 +222,12 @@ func (s *Service) saveLightClientBootstrap(cfg *postBlockProcessConfig) {
 	blockRoot := cfg.roblock.Root()
 	bootstrap, err := lightclient.CreateLightClientBootstrap(cfg.ctx, s.CurrentSlot(), cfg.postState, cfg.roblock)
 	if err != nil {
-		log.WithError(err).Error("Could not create light client bootstrap")
+		log.WithError(err).Error("Saving light client bootstrap failed: Could not create light client bootstrap")
 		return
 	}
 	err = s.cfg.BeaconDB.SaveLightClientBootstrap(cfg.ctx, blockRoot[:], bootstrap)
 	if err != nil {
-		log.WithError(err).Error("Could not save light client bootstrap")
+		log.WithError(err).Error("Saving light client bootstrap failed: Could not save light client bootstrap in DB")
 	}
 }
