+++ consensus-types/interfaces/light_client.go
@@ -31,8 +31,8 @@ type LightClientBootstrap interface {
 	CurrentSyncCommittee() *pb.SyncCommittee
 	SetCurrentSyncCommittee(sc *pb.SyncCommittee) error
 	CurrentSyncCommitteeBranch() (LightClientSyncCommitteeBranch, error)
-	SetCurrentSyncCommitteeBranch(branch [][]byte) error
 	CurrentSyncCommitteeBranchElectra() (LightClientSyncCommitteeBranchElectra, error)
+	SetCurrentSyncCommitteeBranch(branch [][]byte) error
 }
 
 type LightClientUpdate interface {
